TaskManagement.Core/Repository/Impl/ProjectRepository.cs
using Microsoft.EntityFrameworkCore;
using TaskManagement.Core.Data;
using TaskManagement.Core.Models;
using TaskManagement.Core.Models.Enums;

namespace TaskManagement.Core.Repository;

    public class ProjectRepository : IProjectRepository
    {
        private readonly ApplicationDbContext _context;
        public ProjectRepository(ApplicationDbContext context)
        {
            _context = context;
        }

        public async Task<IEnumerable<Project>> GetAllProjectsAsync()
        {
            return await _context.Projects
                .Include(p => p.Tasks)
                .ToListAsync();
        }

        public async Task<Project> GetProjectByIdAsync(int projectId)
        {
            return await _context.Projects
                .Include(p => p.Tasks)
                .FirstOrDefaultAsync(p => p.Id == projectId);
        }

        public async Task<Project> AddProjectAsync(Project project)
        {
            _context.Projects.Add(project);
            await _context.SaveChangesAsync();
            return project;
        }

        public async Task<bool> RemoveProjectAsync(int projectId)
        {
            var project = await GetProjectByIdAsync(projectId);
            if (project == null)
                return false;

            _context.Projects.Remove(project);
            await _context.SaveChangesAsync();
            return true;
        }

        public async Task<bool> HasPendingTasksAsync(int projectId)
        {
            return await _context.TaskItems.AnyAsync(t => t.ProjectId == projectId && t.Status != (int)TaskItemStatus.Concluida);
        }
    }


